Handle ICMP identifier as a 32 bit integer
Currently the ICMP identifier in a flow key is a 16-bit integer,
which determines some mismatches when comparing the flow key
value with the conntrack value, and also displays ICMP identifiers
higher than 0x7FFF as negative integers.

With this change the ICMP identifier in the flow key becomes a
32 bit integer, and is converted appropriately when read from or
written to the ICMP packet.

Fixes: MI-910

(cherry picked from commit 15b8051d20929a63787f6b113b8aa1b31d94bfd6)

+++ midolman/src/main/scala/org/midonet/midolman/monitoring/BinaryFlowRecorder.scala
@@ -21,7 +21,7 @@ import java.util.{ArrayList, List, UUID}
 
 import uk.co.real_logic.sbe.codec.java._
 
-import org.midonet.cluster.flowhistory.{proto, ActionEncoder, BinarySerialization}
+import org.midonet.cluster.flowhistory.{ActionEncoder, BinarySerialization}
 import org.midonet.cluster.flowhistory.proto.{SimulationResult => SbeSimResult,
                                               RuleResult => SbeRuleResult,
                                               DeviceType => SbeDeviceType,
@@ -317,7 +317,7 @@ class BinaryFlowRecorder(val hostId: UUID, config: FlowHistoryConfig)
                             actionEnc.etherType(k.etherType)
                         case k: FlowKeyICMPEcho =>
                             actionEnc.icmpEcho(k.icmp_type,
-                                               k.icmp_code, k.icmp_id)
+                                               k.icmp_code, k.icmp_id.toShort)
                         case k: FlowKeyICMPError =>
                             actionEnc.icmpError(k.icmp_type, k.icmp_code,
                                                 k.icmp_data)
